Restructured Folders, Neoforge working
Restructured Fabric and Neoforge to be in their own folders within their directory to prevent file location issues. Neoforge working as expected.
