+++ fabric/src/main/java/com/lemenok/cobblemontrialsedition/fabric/CobblemonTrialsEditionFabric.java
@@ -1,0 +1,79 @@
+package com.lemenok.cobblemontrialsedition.fabric;
+
+import com.lemenok.cobblemontrialsedition.fabric.block.ModBlocks;
+import com.lemenok.cobblemontrialsedition.fabric.block.entity.ModBlockEntities;
+import com.lemenok.cobblemontrialsedition.config.StructureProperties;
+import com.lemenok.cobblemontrialsedition.fabric.events.SpawnerReplacementHandler;
+import com.lemenok.cobblemontrialsedition.fabric.item.ModCreativeModeTabs;
+import com.lemenok.cobblemontrialsedition.fabric.particle.ModParticles;
+import com.lemenok.cobblemontrialsedition.fabric.potion.ModPotions;
+import com.lemenok.cobblemontrialsedition.fabric.sound.ModSounds;
+import me.shedaniel.autoconfig.AutoConfig;
+import me.shedaniel.autoconfig.serializer.Toml4jConfigSerializer;
+import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.event.lifecycle.v1.ServerChunkEvents;
+import net.fabricmc.fabric.api.event.registry.DynamicRegistries;
+import net.fabricmc.fabric.api.registry.FabricBrewingRecipeRegistryBuilder;
+import net.minecraft.core.Registry;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.item.alchemy.Potions;
+import net.minecraft.world.item.crafting.Ingredient;
+import net.minecraft.world.level.storage.loot.LootTable;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class CobblemonTrialsEditionFabric implements ModInitializer {
+    public static final String MODID = "cobblemontrialsedition";
+    public static final Logger LOGGER = LoggerFactory.getLogger(MODID);
+    public static final SpawnerReplacementHandler SPAWNER_REPLACEMENT_HANDLER = new SpawnerReplacementHandler();
+
+    public static final ResourceKey<Registry<StructureProperties>> COBBLEMON_TRIALS_STRUCTURE_REGISTRY =
+            ResourceKey.createRegistryKey(
+                    ResourceLocation.fromNamespaceAndPath(MODID, "structures")
+            );
+
+    public static final ResourceKey<Registry<StructureProperties>> COBBLEMON_TRIALS_FEATURES_REGISTRY =
+            ResourceKey.createRegistryKey(
+                    ResourceLocation.fromNamespaceAndPath(MODID, "features")
+            );
+
+    public static final ResourceKey<Registry<StructureProperties>> COBBLEMON_TRIALS_DEFAULT_STRUCTURE_REGISTRY =
+            ResourceKey.createRegistryKey(
+                    ResourceLocation.fromNamespaceAndPath(MODID, "defaults")
+            );
+
+    public static final ResourceKey<Registry<LootTable>> COBBLEMON_TRIALS_LOOT_TABLE_REGISTRY =
+            ResourceKey.createRegistryKey(
+                    ResourceLocation.fromNamespaceAndPath(MODID, "loot_table")
+            );
+
+
+    @Override
+    public void onInitialize() {
+        ModBlocks.registerModBlocks();
+        ModBlockEntities.registerBlockEntities();
+        ModParticles.registerParticles();
+        ModPotions.registerPotions();
+        ModSounds.registerSounds();
+        ModCreativeModeTabs.registerItemGroups();
+
+        AutoConfig.register(Config.class, Toml4jConfigSerializer::new);
+
+        // Register Potion Recipe
+        FabricBrewingRecipeRegistryBuilder.BUILD.register(builder -> {
+            builder.registerPotionRecipe(Potions.AWKWARD, Ingredient.of(Items.SCULK), ModPotions.TRIAL_POTION);
+        });
+
+        // Register Datapacks
+        DynamicRegistries.register(COBBLEMON_TRIALS_STRUCTURE_REGISTRY, StructureProperties.CODEC);
+        DynamicRegistries.register(COBBLEMON_TRIALS_FEATURES_REGISTRY, StructureProperties.CODEC);
+        DynamicRegistries.register(COBBLEMON_TRIALS_DEFAULT_STRUCTURE_REGISTRY, StructureProperties.CODEC);
+        DynamicRegistries.register(COBBLEMON_TRIALS_LOOT_TABLE_REGISTRY, LootTable.DIRECT_CODEC);
+
+        // Handle Chunk Load Events
+        ServerChunkEvents.CHUNK_LOAD.register((world, chunk) ->
+                world.getServer().execute(() -> SPAWNER_REPLACEMENT_HANDLER.processNewChunk(world, chunk)));
+    }
+}
