+++ fabric/src/main/java/com/lemenok/cobblemontrialsedition/fabric/particle/ModParticles.java
@@ -1,0 +1,23 @@
+package com.lemenok.cobblemontrialsedition.fabric.particle;
+
+import com.lemenok.cobblemontrialsedition.fabric.CobblemonTrialsEditionFabric;
+import net.fabricmc.fabric.api.particle.v1.FabricParticleTypes;
+import net.minecraft.core.Registry;
+import net.minecraft.core.particles.SimpleParticleType;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.resources.ResourceLocation;
+
+
+public class ModParticles {
+
+    public static final SimpleParticleType UNOWN_PARTICLES =
+            registerParticle("unown_particles", FabricParticleTypes.simple());
+
+    private static SimpleParticleType registerParticle(String name, SimpleParticleType particleType) {
+        return Registry.register(BuiltInRegistries.PARTICLE_TYPE, ResourceLocation.fromNamespaceAndPath(CobblemonTrialsEditionFabric.MODID, name), particleType);
+    }
+
+    public static void registerParticles() {
+        CobblemonTrialsEditionFabric.LOGGER.info("Registering Particles for " + CobblemonTrialsEditionFabric.MODID);
+    }
+}
