+++ fabric/src/main/resources/META-INF/services/com.lemenok.cobblemontrialsedition.integrations.IModIntegrations
@@ -1,1 +1,1 @@
-com.lemenok.cobblemontrialsedition.integrations.FabricModIntegrations
+com.lemenok.cobblemontrialsedition.fabric.integrations.ModIntegrations
