+++ fabric/src/main/resources/fabric.mod.json
@@ -12,13 +12,13 @@
   "environment": "*",
   "entrypoints": {
     "main": [
-      "com.lemenok.cobblemontrialsedition.CobblemonTrialsEditionFabric"
+      "com.lemenok.cobblemontrialsedition.fabric.CobblemonTrialsEditionFabric"
     ],
     "client": [
-      "com.lemenok.cobblemontrialsedition.CobblemonTrialsEditionClientFabric"
+      "com.lemenok.cobblemontrialsedition.fabric.CobblemonTrialsEditionClientFabric"
     ],
     "modmenu": [
-      "com.lemenok.cobblemontrialsedition.ModMenuIntegration"
+      "com.lemenok.cobblemontrialsedition.fabric.ModMenuIntegration"
     ]
   },
   "depends": {
