--- neoforge/src/main/java/com/lemenok/cobblemontrialsedition/events/ModEvents.java
@@ -1,21 +1,0 @@
-package com.lemenok.cobblemontrialsedition.events;
-
-import com.lemenok.cobblemontrialsedition.CobblemonTrialsEdition;
-import com.lemenok.cobblemontrialsedition.potion.ModPotions;
-import net.minecraft.world.item.Items;
-import net.minecraft.world.item.alchemy.PotionBrewing;
-import net.minecraft.world.item.alchemy.Potions;
-import net.neoforged.bus.api.SubscribeEvent;
-import net.neoforged.fml.common.EventBusSubscriber;
-import net.neoforged.neoforge.event.brewing.RegisterBrewingRecipesEvent;
-
-@EventBusSubscriber(modid = CobblemonTrialsEdition.MODID)
-public class ModEvents {
-
-    @SubscribeEvent
-    public static void onBrewingRecipeRegister(RegisterBrewingRecipesEvent event) {
-        PotionBrewing.Builder builder = event.getBuilder();
-
-        builder.addMix(Potions.AWKWARD, Items.SCULK, ModPotions.TRIAL_POTION);
-    }
-}
