+++ neoforge/src/main/java/com/lemenok/cobblemontrialsedition/neoforge/events/ClientModHandler.java
@@ -1,0 +1,26 @@
+package com.lemenok.cobblemontrialsedition.neoforge.events;
+
+import com.lemenok.cobblemontrialsedition.neoforge.CobblemonTrialsEdition;
+import com.lemenok.cobblemontrialsedition.neoforge.block.entity.ModBlockEntities;
+import com.lemenok.cobblemontrialsedition.client.renderer.blockentity.CobblemonTrialSpawnerRenderer;
+import com.lemenok.cobblemontrialsedition.neoforge.particle.ModParticles;
+import com.lemenok.cobblemontrialsedition.particle.UnownParticles;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.common.EventBusSubscriber;
+import net.neoforged.neoforge.client.event.EntityRenderersEvent;
+import net.neoforged.neoforge.client.event.RegisterParticleProvidersEvent;
+
+@EventBusSubscriber(modid = CobblemonTrialsEdition.MODID, value = Dist.CLIENT)
+public class ClientModHandler {
+
+    @SubscribeEvent
+    public static void registerRenderers(EntityRenderersEvent.RegisterRenderers event) {
+        event.registerBlockEntityRenderer(ModBlockEntities.COBBLEMON_TRIAL_SPAWNER.get(), CobblemonTrialSpawnerRenderer::new);
+    }
+
+    @SubscribeEvent
+    public static void registerParticleFactories(RegisterParticleProvidersEvent event) {
+        event.registerSpriteSet(ModParticles.UNOWN_PARTICLES.get(), UnownParticles.Provider::new);
+    }
+}
