+++ neoforge/src/main/resources/META-INF/services/com.lemenok.cobblemontrialsedition.integrations.IModIntegrations
@@ -1,0 +1,1 @@
+com.lemenok.cobblemontrialsedition.neoforge.integrations.ModIntegrations
